requirements.txt
Django==3.1
psycopg2-binary==2.8.5
Pillow==7.1
django-ckeditor
django-crispy-forms
Celery==4.4.7
flower==0.9.5
braintree==4.3.0
setuptools==50.1.0
cffi==1.14.2
html5lib==1.1
cairocffi==1.1.0
tinycss2==1.0.2
cssselect2==0.3.0
CairoSVG==2.4.2
Pyphen==0.9.5
WeasyPrint
gunicorn
python-decouple
dj-database-url
whitenoise